things: split api into SQL/JXA paths with useUnofficialApi toggle
- Split api.ts into api-sql.ts (SQLite reads) and api-jxa.ts (Apple Events)
- api.ts is now a thin facade switching on the useUnofficialApi preference
- Add SQL reads for all built-in lists including logbook and trash
- Match SQL list output to JXA (order and content verified via compare tool)
- Add compare-jxa-sql dev command for parity verification
- Remove dead code and unused exports across the extension
